feat: calculate fees for multiple redemptions at once
This is going to be useful for both `redeemMany` and
eventually `redeemAuto` in the pool contract as these
functions already support redemption of multiple TCO2s
within a single transaction.

We can even get rid of `calculateRedemptionFees` altogether
since I don't forsee it being used in the Toucan pools
but this can be done in a follow-up if needed.