Break down fee calculator interfaces
The pool contract needs to estimate the fee
to charge for a deposit or redemption separately
from executing the distribution of the total fee
to the fee recipients because in between the two
the pool contract needs to either burn the tokens
from the user in case of a redemption or mint tokens
to the user in case of a deposit.

In order to best facilitate this requirement the
current fee calculator interface needs to separate
estimating the fee of a deposit/redemption from
estimating the shares for each fee recipient.