Pin contract versions
This is a recommended practice by security
auditors, also prb-math requires 0.8.19 version
so pin the main fee calculator contract to 0.8.19.

This change aligns the OZ contract version with
the version used in the rest of the Toucan protocols
which will reduce the effort we will need to spend
during auditing.