chore(sol): do not throw on zero fee returned
There can be edge cases in the pool where the pool
may provide one or more wei in the calculation
functions and the calculation functions return zero.

For example, this can happen when CHAR is using a
new feature called `minimumTCLSeedingThreshold` that:
1. avoids charging fees to deposits that keep the total
   underlying supply below a certain threshold
2. charges fees partially when the deposit brings the
   total underlying supply above the threshold
3. charges fees normally above the threshold

Given a deposit that falls under 2, it may be that
the chargeable amount is one or more few wei. In this
scenario, the pool currently has to handle an error
returned by the fee calculator which decreases devexp
and complicates pool code.

On the other hand, the pool code is already written
to loop through the returned arrays that are part of
the FeeDistribution struct so returning a struct with
empty arrays means this edge case can be handled in
the pool seamlessly.